Travis CI build update
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@39368 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/build/msys2/install_deps.sh
@@ -16,11 +16,14 @@ pacman --sync --noconfirm --needed  \
 
 XA_VERSION=2.3.11
 
-cd /usr/local
-mkdir -p src
-cd src
-wget http://www.floodgap.com/retrotech/xa/dists/xa-${XA_VERSION}.tar.gz
-tar -xzf xa-${XA_VERSION}.tar.gz
-cd xa-${XA_VERSION}
-make mingw install
-cp /usr/local/bin/xa.exe /usr/local/bin/xa65.exe
+if [ ! -e /usr/local/bin/xa65.exe ]
+then
+    cd /usr/local
+    mkdir -p src
+    cd src
+    wget http://www.floodgap.com/retrotech/xa/dists/xa-${XA_VERSION}.tar.gz
+    tar -xzf xa-${XA_VERSION}.tar.gz
+    cd xa-${XA_VERSION}
+    make mingw install
+    cp /usr/local/bin/xa.exe /usr/local/bin/xa65.exe
+fi
